[v2] refactor(tests): shared compilation helpers live in tests/support
