+++ crates/solidity-v2/outputs/cargo/slang_solidity/src/tests/create.rs
@@ -1,42 +1,5 @@
-use slang_solidity_v2_common::evm_targets::EvmTarget;
-
-use crate::compilation::{CompilationUnit, Configuration, FileId, ImportResolver};
+use super::support::{compile, contract};
 use crate::diagnostics::DiagnosticExtensions;
-use crate::diagnostics::kinds::compilation::UnresolvedImport;
-use crate::utils::LanguageVersion;
-
-/// Resolves every import path to a file of the same name.
-struct TestImportResolver;
-
-impl ImportResolver for TestImportResolver {
-    fn resolve_import(
-        &mut self,
-        _source_file_id: &FileId,
-        import_path: &str,
-    ) -> Result<FileId, UnresolvedImport> {
-        Ok(import_path.into())
-    }
-}
-
-fn compile<'s>(sources: impl IntoIterator<Item = (FileId, &'s str)>) -> CompilationUnit {
-    CompilationUnit::create(Configuration {
-        language_version: LanguageVersion::LATEST,
-        evm_target: EvmTarget::LATEST,
-        sources,
-        resolver: TestImportResolver,
-    })
-}
-
-fn contract(name: &str, imports: &[&str]) -> String {
-    use std::fmt::Write;
-
-    let imports = imports.iter().fold(String::new(), |mut text, path| {
-        writeln!(text, "import \"{path}\";").unwrap();
-        text
-    });
-
-    format!("pragma solidity ^0.8.0;\n{imports}\ncontract {name} {{}}\n")
-}
 
 #[test]
 fn compiles_every_source_it_is_given() {
