+++ crates/solidity-v2/outputs/cargo/slang_solidity/src/tests/mod.rs
@@ -3,5 +3,6 @@ mod ast;
 mod create;
 mod diagnostics;
 mod fixtures;
+mod support;
 mod thread_safety;
 mod unit;
